Add files via upload
The code performs web scraping on the Politifact website to obtain statements related to climate change and label them according to their truthfulness. To do this, the first 12 pages of the fact-checking section of the website are traversed and the statement, evidence, and truthfulness label of each statement are extracted.

On each page, a web request is made to the page's URL and BeautifulSoup is used to parse the page content. The statement's date is sought, and the statement itself is extracted. Then the page of the statement is accessed to extract all the evidence.

The text of the evidence is divided into sentences, and the last sentence containing a truthfulness label is sought. A truthfulness label is assigned to the statement based on the content of that last sentence.

If a truthfulness label is not found in the last sentence, an image with a truthfulness label is searched for, and the corresponding label is assigned. If a truthfulness label is not found anywhere, None is assigned.

Finally, a dataframe is created with the statements, evidence, truthfulness labels, and dates, and exported as a CSV file.