Add files via upload
With the 604 claims we have (from Nigeria and Zambia elections), I have used the GPT-3 API to generate 5 true claims and 5 false claims based on the available evidence, resulting in a total of 6184 claims (the quality of these tweets, which is limited by the GPT API, needs to be checked). In addition, I am creating a structured dataset that includes the claim, the top 5 evidences with S-BERT, and the label.

I have also incorporated predictions from various models to enrich the dataset:

I used the 'facebook/bart-large-mnli' model with zero-shot-classification to make an initial prediction of True or False.

With the 'arpanghoshal/EmoRoBERTa' model, 28 different types of emotions are detected, and I included the probabilities for each emotion (scores).

Additionally, I used another model called 'unitary/unbiased-toxic-roberta' to detect types of toxicity or biases in the claims, such as toxicity, insults, identity attacks, etc.

I also used zero-shot-classification to categorize the claims into election-related categories, such as voter registration, electoral administration, electoral violence, political parties, campaign strategies, election monitoring, electoral transparency, and many more.  

I have also added a frequency matrix of up to trigrams of proper names both during the S-BERT process and in the training of the models. This seems to help find evidence that is more relevant to the claim. I have created a stacking classifier with the four models, adding an XGBoost meta-classifier that combines the probabilities for making the final prediction.